+++ doc/scripts/build_par_rst.py
@@ -15,7 +15,8 @@ import numpy
 from pkg_resources import resource_filename
 from pypeit.par import pypeitpar
 from pypeit.par.parset import ParSet
-from pypeit.spectrographs.util import load_spectrograph, valid_spectrographs
+from pypeit.spectrographs.util import load_spectrograph
+from pypeit.defs import pypeit_spectrographs
 
 #-----------------------------------------------------------------------------
 #def class_name(p):
@@ -81,8 +82,7 @@ if __name__ == '__main__':
                            ':class:`pypeit.spectrographs.spectrograph.Spectrograph` class.', 72)
     lines += ['']
 
-    spectrographs = valid_spectrographs()
-    for spec in spectrographs:
+    for spec in pypeit_spectrographs:
         s = load_spectrograph(spec)
         lines += [ ' '.join([s.telescope['name'], s.camera, '(``{0}``)'.format(s.spectrograph)]) ]
         lines += [ '-'*len(lines[-1]) ]