+++ pypeit/scripts/launch_identify.py
@@ -7,6 +7,8 @@
 Launch the identify GUI tool.
 """
 
+from IPython import embed
+
 
 def parser(options=None):
     import argparse
@@ -17,17 +19,17 @@ def parser(options=None):
                                                  'Run above the Masters/ folder',
                                      formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 
-    parser.add_argument('file', type=str, default=None, help='PypeIt MasterArc file')
-    parser.add_argument("--lamps", default=None, help="Comma separated list of calibration lamps (no spaces)",
-                        action="store_true")
+    parser.add_argument('arc_file', type=str, default=None, help='PypeIt MasterArc file')
+    parser.add_argument('slits_file', type=str, default=None, help='PypeIt MasterSlits file')
+    parser.add_argument("--lamps", type=str, help="Comma separated list of calibration lamps (no spaces)")
     parser.add_argument("--wmin", default=3000.0, help="Minimum wavelength range",
                         action="store_true")
     parser.add_argument("--wmax", default=10000.0, help="Maximum wavelength range",
                         action="store_true")
     parser.add_argument("--slit", default=0, help="Slit number to wavelength calibrate",
                         action="store_true")
-    parser.add_argument("--det", default=1, help="Detector index",
-                        action="store_true")
+    parser.add_argument("--det", default=1, help="Detector index", action="store_true")
+    parser.add_argument("--rmstol", default=0.1, help="RMS tolerance", action="store_true")
 
     return parser.parse_args() if options is None else parser.parse_args(options)
 
@@ -37,25 +39,25 @@ def main(args):
     import os
     import sys
     import astropy.io.fits as fits
-    from pypeit.masterframe import MasterFrame
+    from pypeit import masterframe
     from pypeit.spectrographs.util import load_spectrograph
     from pypeit.core import parse
     from pypeit.core import gui
     from pypeit.core.wavecal import waveio, templates
     from pypeit.wavecalib import WaveCalib
     from pypeit import slittrace
-    from pypeit.images import pypeitimage
+    from pypeit.images.buildimage import ArcImage
 
     # Load the MasterArc file
-    if os.path.exists(args.file):
-        arcfil = args.file
+    if os.path.exists(args.arc_file):
+        arcfil = args.arc_file
     else:
         try:
-            arcfil = "Masters/{0:s}".format(args.file)
+            arcfil = "Masters/{0:s}".format(args.arc_file)
         except FileNotFoundError:
             print("Could not find MasterArc file.")
             sys.exit()
-    msarc = pypeitimage.PypeItImage.from_file(arcfil)
+    msarc = ArcImage.from_file(arcfil)
 
     mdir = msarc.head0['MSTRDIR']
     mkey = msarc.head0['MSTRKEY']
@@ -76,20 +78,23 @@ def main(args):
     par['lamps'] = lamplist
 
     # Load the slits
-    slits = slittrace.SlitTraceSet.from_master(mkey, mdir)
+    slits = slittrace.SlitTraceSet.from_file(args.slits_file)
+    # Reset the mask
+    slits.mask = slits.mask_init
 
     # Check if a solution exists
-    solnname = os.path.join(mdir, MasterFrame.construct_file_name('WaveCalib', mkey, file_format='json'))
+    solnname = os.path.join(mdir, masterframe.construct_file_name(WaveCalib, mkey))
     wv_calib = waveio.load_wavelength_calibration(solnname) if os.path.exists(solnname) else None
 
     # Load the MasterFrame (if it exists and is desired)?
     wavecal = WaveCalib(msarc, slits, spec, par, binspectral=slits.binspec, det=args.det,
-                        master_key=mkey, master_dir=mdir, msbpm=msarc.mask)
+                        master_key=mkey, msbpm=msarc.fullmask)
     arccen, arc_maskslit = wavecal.extract_arcs()
 
     # Launch the identify window
-    arcfitter = gui_identify.initialise(arccen, slit=int(args.slit), par=par, wv_calib_all=wv_calib,
-                                        wavelim=[args.wmin, args.wmax])
+    arcfitter = gui.identify.initialise(arccen, slit=int(args.slit), par=par, wv_calib_all=wv_calib,
+                                        wavelim=[args.wmin, args.wmax],
+                                        nonlinear_counts=spec.nonlinear_counts(msarc.detector))
     final_fit = arcfitter.get_results()
 
     # Ask the user if they wish to store the result in PypeIt calibrations
@@ -100,7 +105,7 @@ def main(args):
         if ans == 'y':
             gratname = fits.getheader(msarc.head0['F1'])[spec.meta['dispname']['card']].replace("/", "_")
             dispangl = "UNKNOWN"
-            templates.pypeit_identify_record(final_fit, binspec, specname, gratname, dispangl)
+            templates.pypeit_identify_record(final_fit, slits.binspec, specname, gratname, dispangl)
             print("Your wavelength solution has been stored")
             print("Please consider sending your solution to the PypeIt team!")
     else: