+++ pypeit/spectrographs/gemini_gmos.py
@@ -52,16 +52,6 @@ class GeminiGMOSSpectrograph(spectrograph.Spectrograph):
 
         self.meta['datasec'] = dict(ext=1, card='DATASEC')
 
-    def configuration_keys(self):
-        """
-        Extra keys for defining the configuration
-
-        Returns:
-
-        """
-        return ['datasec']
-
-
     def compound_meta(self, headarr, meta_key):
         """
 
@@ -88,7 +78,7 @@ class GeminiGMOSSpectrograph(spectrograph.Spectrograph):
             #& (fitstbl['idname'] == 'OBJECT')
         if ftype in ['arc', 'tilt']:
             return good_exp & (fitstbl['target'] == 'CuAr')#& (fitstbl['idname'] == 'ARC')
-        if ftype in ['pixelflat', 'trace']:
+        if ftype in ['pixelflat', 'trace', 'illumflat']:
             return good_exp & (fitstbl['target'] == 'GCALflat')#& (fitstbl['idname'] == 'FLAT')
         if ftype == 'bias':
             return good_exp & (fitstbl['target'] == 'Bias')#& (fitstbl['idname'] == 'BIAS')
@@ -102,6 +92,9 @@ class GeminiGMOSSpectrograph(spectrograph.Spectrograph):
         Set default parameters for Gemini GMOS reductions.
         """
         par = pypeitpar.PypeItPar()
+
+
+
         par['calibrations']['slitedges']['edge_thresh'] = 20.
         par['calibrations']['slitedges']['fit_order'] = 3
 
@@ -181,7 +174,7 @@ class GeminiGMOSSpectrograph(spectrograph.Spectrograph):
         """
         cfg_keys = super(GeminiGMOSSpectrograph, self).configuration_keys()
         # Add grating tilt
-        return cfg_keys+['dispangle']
+        return cfg_keys+['dispangle', 'datasec']
 
     def get_rawimage(self, raw_file, det):
         """
@@ -596,8 +589,8 @@ class GeminiGMOSNHamSpectrograph(GeminiGMOSNSpectrograph):
             par['calibrations']['wavelengths']['reid_arxiv'] = 'gemini_gmos_r400_ham.fits'
         elif self.get_meta_value(scifile, 'dispname')[0:4] == 'B600':
             par['calibrations']['wavelengths']['reid_arxiv'] = 'gemini_gmos_b600_ham.fits'
-        #
         return par
+
 
 class GeminiGMOSNE2VSpectrograph(GeminiGMOSNSpectrograph):
     """