+++ pypeit/spectrographs/wht_isis.py
@@ -147,13 +147,19 @@ class WHTISISBlueSpectrograph(WHTISISSpectrograph):
         # Ignore PCA
         par['calibrations']['slitedges']['sync_predict'] = 'nearest'
 
+
+        # JFH Is this correct?
+        # Processing steps
+        turn_off = dict(use_overscan=False)
+        par.reset_all_processimages_par(**turn_off)
+
         # Turn off the overscan
-        for ftype in par['calibrations'].keys():
-            try:
-                par['calibrations'][ftype]['process']['overscan'] = 'none'
-            except (TypeError, KeyError):
-                pass
-        par['scienceframe']['process']['overscan'] = 'none'
+        #for ftype in par['calibrations'].keys():
+        #    try:
+        #        par['calibrations'][ftype]['process']['overscan'] = 'none'
+        #    except (TypeError, KeyError):
+        #        pass
+        par['scienceframe']['process']['use_overscan'] = False
         # Make a bad pixel mask
         par['calibrations']['bpm_usebias'] = True
         # Set pixel flat combination method
@@ -292,7 +298,7 @@ class WHTISISRedSpectrograph(WHTISISSpectrograph):
                 par['calibrations'][ftype]['process']['overscan'] = 'none'
             except (TypeError, KeyError):
                 pass
-        par['scienceframe']['process']['overscan'] = 'none'
+        par['scienceframe']['process']['use_overscan'] = False
         # Make a bad pixel mask
         par['calibrations']['bpm_usebias'] = True
         # Set pixel flat combination method